django-resume-builder (4)/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
# psycopg2-binary==2.9.7
django-environ==0.11.2
django-filter==23.3
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5
Pillow==10.0.1
django-allauth==0.57.0
reportlab==4.0.4
python-docx==0.8.11
# weasyprint==60.2
django-wkhtmltopdf==3.4.0
